+++ src/shared/lib/message/prepare-last-message-for-conversation-list/config.interface.ts
@@ -2,9 +2,9 @@ import type { Dto } from '@specialized-messenger/api/specs';
 
 export type MessageTypeName = Dto['Message']['type']['name'];
 
-export type Preparer = (
-  message: Dto['Message'],
-  participants: Dto['ConversationParticipant'][],
-) => string;
+export type Preparer = (data: {
+  lastMessage: Dto['Message'];
+  conversation: Dto['Conversation'];
+}) => string;
 
 export type Config = Record<MessageTypeName, Preparer>;