+++ src/shared/lib/message/prepare-last-message-for-conversation-list/units/system-conversation-created.ts
@@ -2,8 +2,11 @@ import { isSystemConversationCreatedMessage } from '../../message-checker/is-sys
 import { MESSAGE_PARSE_ERROR } from '../../message-parse-error';
 import type { Preparer } from '../config.interface';
 
-export const prepareSystemConversationCreatedMessage: Preparer = (message) => {
-  if (!isSystemConversationCreatedMessage(message)) return MESSAGE_PARSE_ERROR;
+export const prepareSystemConversationCreatedMessage: Preparer = ({
+  lastMessage,
+}) => {
+  if (!isSystemConversationCreatedMessage(lastMessage))
+    return MESSAGE_PARSE_ERROR;
 
   return 'Чат создан';
 };