+++ src/shared/lib/message/prepare-last-message-for-conversation-list/units/system-user-left.ts
@@ -5,14 +5,16 @@ import { isSystemUserLeftMessage } from '../../message-checker/is-system-user-le
 import { MESSAGE_PARSE_ERROR } from '../../message-parse-error';
 import type { Preparer } from '../config.interface';
 
-export const prepareSystemUserLeftMessage: Preparer = (
-  message,
-  participants,
-) => {
-  if (!isSystemUserLeftMessage(message)) return MESSAGE_PARSE_ERROR;
+export const prepareSystemUserLeftMessage: Preparer = ({
+  lastMessage,
+  conversation,
+}) => {
+  if (!isSystemUserLeftMessage(lastMessage)) return MESSAGE_PARSE_ERROR;
 
-  const leftParticipant = participants.find(
-    (participant) => participant.user.publicId === message.content.userPublicId,
+  // Т.к. это ПОСЛЕДНЕЕ сообщение чата, то после него пользователь находится гарантированно в removedParticipants
+  const leftParticipant = conversation.removedParticipants.find(
+    (participant) =>
+      participant.user.publicId === lastMessage.content.userPublicId,
   );
 
   return `Пользователь ${leftParticipant ? getUserFullName(leftParticipant.user) : UNDEFINED_USER} покинул чат`;