+++ src/shared/lib/message/prepare-last-message-for-conversation-list/units/text.ts
@@ -1,8 +1,8 @@
 import { isTextMessage } from '../../message-checker/is-text-message';
 import type { Preparer } from '../config.interface';
 
-export const prepareTextMessage: Preparer = (message) => {
-  if (!isTextMessage(message)) return '';
+export const prepareTextMessage: Preparer = ({ lastMessage }) => {
+  if (!isTextMessage(lastMessage)) return '';
 
-  return `${message.author.firstName}: ${message.content.text}`;
+  return `${lastMessage.author.firstName}: ${lastMessage.content.text}`;
 };