+++ src/shared/lib/message/prepare-message-for-active-conversation/config.interface.ts
@@ -2,9 +2,9 @@ import type { Dto } from '@specialized-messenger/api/specs';
 
 export type MessageTypeName = Dto['Message']['type']['name'];
 
-export type Preparer = (
-  message: Dto['Message'],
-  conversation: Dto['Conversation'],
-) => string;
+export type Preparer = (data: {
+  message: Dto['Message'];
+  allParticipants: Dto['ConversationParticipant'][];
+}) => string;
 
 export type Config = Record<MessageTypeName, Preparer>;