+++ src/shared/lib/message/prepare-message-for-active-conversation/units/system-conversation-created.ts
@@ -4,7 +4,9 @@ import { isSystemConversationCreatedMessage } from '../../message-checker/is-sys
 import { MESSAGE_PARSE_ERROR } from '../../message-parse-error';
 import type { Preparer } from '../config.interface';
 
-export const prepareSystemConversationCreatedMessage: Preparer = (message) => {
+export const prepareSystemConversationCreatedMessage: Preparer = ({
+  message,
+}) => {
   if (!isSystemConversationCreatedMessage(message)) return MESSAGE_PARSE_ERROR;
 
   return `Чат создан в ${getLocalTime(new Date(message.createdAt))}`;